fix: inconsistent followers and ghosts lookup
every week, when people — well, mostly me — get weekly digest, it is inconsitent.

for instance, say i have three people who unfollowed me, i'll only get a mail saying one person did, and vice versa when the follower count increases too.

this problem was as a result of computing the increment or decrement just in time when the snapshot was taken, and because the SQL query was unncessarily complex where i was trying to group the list by timestamps which sometimescan be very inaccurate.

relationships aren't all that bad.

eventually, i had to resort to having a separate table for ghosts with a foreign key in relationship with a user, so now the weekly summary job basically compares both tables in a week epoch.